Add files via upload
This text file contains a record of user responses to prompts for user input made by the BIGS908 software application. The name of the file includes a set of characters consisting of "J23h25s6dgame13B". The set of characters refers to the following.  "J" indicates that each deck includes two jokers. "23h25s" indicates that the strategy for the play of the dealer's hand calls for the dealer to stand on hard 23 and soft 25. 6d indicates that the dealer deals from a six deck shoe. "game13" is the same game as is found in the "Finnish-27-907docs" repository. "B" indicates this is the second game simulation conducted using the predetermined set of game rules specified for game13.