Add files via upload
This is a rather lengthy text file saved by the BIGS908 software application. It contains a record of the results of a simulation of a blackjack-like game . The game belongs to the Finnish 27 family of table-card games. A text file named 'game13i.txt' contains a record of the rules of the game. The text file can be found in the Finnish-27-907docs repository. A text file named 'J23h25s6dgame13i.txt' contains a record of the program parameters specified for this game simulation. The program parameters indicate the following. The game simulation consists of one billion rounds of play divided evenly into one hundred sessions of ten-million round of play each. At the end of each session, the BIGS908 software application displays to the application window various statistics. The statistics are about the last hand played in the session and about the outcome of the entire session. The BIGS908 software application saved a record of the application window to the text file named 'J23h25s6dgame13Bo.