Add files via upload
This text file was saved by the BIGS908 software application. It contains a record of the statistics gathered during the course of a game simulation. The game simulation was of a blackjack-like game belonging to the Finnish 27 family of table card games. A text file named 'game13i.txt' contains a record of the game rules. A text file named 'game13o.txt' contains a set of combination-dependent-playing-strategy tables found for the game. Said text files can be found in the Finnish-27-907docs repository. The user loaded data about the set of combination-dependent-playing-strategy tables into program memory of the BIGS908 software application. The BIGS908 software application used this data to program a computer-controlled player to make decisions about how to play the players hand.