Add files via upload
This text file contains a record of user responses to prompts for user input made by the BIGS908 software application. The name of the file includes a set of characters consisting of "J23h25s6dgame13TDA5". The set of characters refers to the following.  "J" indicates that each deck includes two jokers. "23h25s" indicates that the strategy for the play of the dealer's hand calls for the dealer to stand on hard 23 and soft 25. 6d indicates that the dealer deals from a six deck shoe. "game13" is found in the "Finnish-27-907docs" repository. "TDA" indicates that the playing strategy table used by the computer-controlled player to make decisions on how to play the player's hand is a total-dependent-basic-strategy table. "5" indicates this is the fifth game simulation conducted using the predetermined set of game rules specified for game13.