Add files via upload
This text file was saved by the BIGS908 software application. It contains a record of the statistics gathered during the course of a game simulation. The game simulation was of a blackjack-like game belonging to the Finnish 27 family of table card games. A text file named 'game13i.txt' contains a record of the game rules. A text file named 'game13TDAd.txt' contains a set of combination-dependent-playing-strategy tables found for the game. Said text files can be found in the Finnish-27-907docs repository. The set of combination-dependent-playing-strategy tables is actually a total-dependent-basic-strategy table displayed in the format of a set of combination-dependent-playing-strategy tables. During the game simulation, a computer-controlled player used the set of combination-dependent-playing-strategy tables to make decisions on how to play the player's hand.